+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Detection of NES 2.0 ROM image format.
 - Correct size of PRG/CHR RAM that is read from ROM images (iNES / NES 2.0).
 - MMC1 mapper imlementation (PRG RAM protection, PRG ROM mapping).
+- MMC3 mapper implementation (PRG RAM protection).
 
 ## [0.3.0] - 2015-08-09
 ### Fixed